update favorite state in database when icon is clicked (via API call)
