+++ src/NoteEditor.client.tsx
@@ -14,6 +14,7 @@ import NotePreview from './NotePreview';
 import {useRefresh} from './Cache.client';
 import {useLocation} from './LocationContext.client';
 import {ILocation} from './types';
+import {useMutation} from "./useMutation";
 
 interface NoteEditorProps {
   noteId: number | null;
@@ -159,46 +160,6 @@ const NoteEditor: React.FC<NoteEditorProps> = ({
   );
 };
 
-function useMutation({endpoint, method}: {endpoint: string; method: string}) {
-  const [isSaving, setIsSaving] = useState(false);
-  const [didError, setDidError] = useState(false);
-  const [error, setError] = useState(null);
-  if (didError) {
-    // Let the nearest error boundary handle errors while saving.
-    throw error;
-  }
 
-  async function performMutation(
-    payload: {title?: string; body?: string},
-    requestedLocation: ILocation
-  ) {
-    setIsSaving(true);
-    try {
-      const response = await fetch(
-        `${endpoint}?location=${encodeURIComponent(
-          JSON.stringify(requestedLocation)
-        )}`,
-        {
-          method,
-          body: JSON.stringify(payload),
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        }
-      );
-      if (!response.ok) {
-        throw new Error(await response.text());
-      }
-      return response;
-    } catch (e) {
-      setDidError(true);
-      setError(e as any);
-    } finally {
-      setIsSaving(false);
-    }
-  }
-
-  return {isSaving, performMutation};
-}
 
 export default NoteEditor;
