+++ src/types.ts
@@ -16,7 +16,7 @@ export interface Note {
   id: number;
   title: string;
   body: string;
-  is_favorite: boolean;
+  favorite: boolean;
   updated_at: Date;
 }
